pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pilk"
version = "0.0.1"
description = "PILK — local-first execution OS for agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Aaron Pilk" }]
dependencies = [
    "anthropic>=0.42",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "structlog>=24.4",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
    "websockets>=13",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-multipart>=0.0.12",
    "browserbase>=1.0",
    "playwright>=1.47",
    "google-auth>=2.30",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.140",
    "pyjwt[crypto]>=2.9",
    "pypdf>=5.0",
    "python-docx>=1.1",
    "markdown>=3.10",
    "weasyprint>=68.0",
    "feedparser>=6.0",
    "numpy>=2.4.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
pilkd = "core.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "scripts*"]
exclude = ["tests*", "ui*", "agents*", "docs*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
